header: correctly align baselines when logged in (#6669)
Warning: you won't be able to unsee what I'm about to show you.

When a user is logged out of crates.io, the text `Browse All Crates` is
nicely aligned with `Log In With GitHub`. Awesome.

When a user is logged _in_ to crates.io, the baseline of the user's name
is three pixels higher than the `Browse All Crates` link. Ruinous.

The reason for this is that the `UserAvatar` component ends up being
taller than the `Browse All Crates` element due to the extra height of
the avatar image (22px vs the 16px font size). If we apply a small
negative `margin-top` to the avatar and make all the `.nav` elements
stretch to the same height by using `align-items: stretch`, then both
elements are then rendered with the same baseline.