admin: add a new admin backend to yank and unyank crates
This uses our existing `minijinja` dependency to implement a (mostly)
static HTML admin console that the crates.io team can use to administer
crates without needing direct database access. For now, the only
administrative action allowed is yanking and unyanking crate versions,
but further actions are anticipated to be added in the near future.

The spiciest part of this commit is probably the routing changes, rather
than the actual templating code and controller changes, since these need
to be applied across the development server, nginx, and anything else
that's in front of our frontend and backend servers.