Purge all domains on Fastly
When invalidating the cache on Fastly using the "single URL" endpoint,
the content is only purged for that specific URL. This can cause
inconsistencies, since we have two domains associated with each Fastly
service (namely `static.*` and `fastly-static.*`). The invalidation
logix has been refactor to purge a path for both domains.