tarball: Fail if `Cargo.toml` manifest is missing or invalid
This changes the `crates_io_tarball` package to a) require a `Cargo.toml` file to always be present in a crate file, and b) no longer silently fail if the `Cargo.toml` file can't be parsed.

This ensures that only crate files with valid manifests end up on crates.io. We previously accepted almost arbitrary tarballs, even without any manifests.

Unfortunately this change also required a few changes to our test suite, which previously did not include `Cargo.toml` files in most of the tarballs sent to the server. The `Cargo.toml` that we send now is a veeeery basic manifest and is actually still lacking the required `name` and `version` fields, but, since we don't enforce this part yet, are left out for now.