db: fix placement of ORDER BY for reverse dependencies
Postgres is allowed to reorder the results of every query unless
an ORDER BY is present. Putting the ORDER BY in a subquery then
apparently gave an order to the query, but Postgres was always
allowed to reorder the results.

Also this makes it clearer to the query optimizer what we're
trying to do.