crates_io_tarball: error if multiple manifests are found
This will handle both differences in case (ie a crate file that contains
both `package-x.y.z/Cargo.toml` and `package-x.y.z/cargo.toml`, as may
happen on a case-sensitive filesystem) and malformed tarballs that
contain multiple copies of the same file.

Effort has been made to otherwise preserve the same semantics as `cargo`
where, if there is only one manifest, only `Cargo.toml` and `cargo.toml`
are accepted. (It's debatable if we should even accept `cargo.toml`, and
recent versions of `cargo` won't allow crates to be published with that
casing, but since there may be legacy setups out there that are still
publishing crates with this variation and we've accepted them in the
past, we may as well preserve that behaviour for now.)

Fixes #6978.