Unify `assert_eq!()` argument order
The majority of assertions in this project use `actual, expected` ordering. This changes the ones with `expected, actual` ordering to use the other one.

This makes it more consistent when trying to understand what `left` and `right` mean in the test suite output.

The `claims` assertion library that we use also assumed `actual, expected` ordering.