models/krate: Move URL and name validation out of `create_or_update()`
`create_or_update()` is only called in four places. Three of them are within test code, where a `create()` fn would be sufficient and where we don't necessarily need the same kind of validation. The fourth place is the publish endpoint, where this validation is definitely needed, but could potentially be combined with other validation code in the future.