Replace `ring` dependency with `p256` (#7252)
The `ring` project is releasing new versions without tags on the corresponding repository or even changelog entries, so we basically have no idea of what is happening over there. `p256` is a pure-Rust alternative for the only `ring` algorithm that we're currently using and it also includes functions to parse the PEM format public keys.