Use custom crates.io buildpack
Using this custom buildpack brings our slug size down from 303 MB to 169 MB. It also appears to solve a Rust build caching issue with the generic Rust buildpack that we had been using so far.

The main source of slug size savings is the removal of the `node_modules` folder after the build, and that the Node.js binary is not included in the slug, since it is only needed at build-time.