Heroku: Remove Datadog buildpack
According to the infra team, the buildpack is only needed for additional per-dyno metrics, and the log collection is independent from this. This buildpack however is responsible for a 90 MB increase of the slug size on Heroku. If we don't gain that much from having it, then we should remove it and improve our dyno boot times again :)