Replace `dsl::*` imports
Using these kinds of wildcard imports makes it hard to use variable or argument names that match the names of the table columns. This leads to argument names like `token_` or `id_`. Using the table name prefix is a little more verbose, but also makes it much easier to see where the code is referring to a column vs. a regular variable.