swirl/runner: Make `Runner` generic over `Context`
Instead of assuming that the `Runner` will always work in combination with the `Environment` struct, we can make it generic now and let the job structs define what type of context they assume to get passed into them.

This again, reduces the tight coupling between the `Runner` and our application-specific code (i.e. the `Environment` struct).