middleware/block_traffic: Combine middleware functions
Before this commit the compile time of the `crates_io` package on my machine was around 50 seconds. After this commit it dropped down to 5 seconds. I'm not entirely sure what the root cause of this is, but `git bisect` suggested that the addition of the `block_by_ip` middleware layer was at least partially responsible. It looks like the more middleware layers we add, the longer the app takes to compile. It might also be caused by the `from_fn_with_state()` fn usage, which this commit reduces a bit... tl;dr don't know, but it's faster now 🤷‍♂️