bin/background-worker: Remove `build_runner()` fn
The `Runner::new()` fn itself doesn't do much, so re-running it should not change the outcome compared to re-running `runner.run_all_pending_jobs()` directly.

The only difference is the re-initialization of the r2d2 database connection pool. The database pool should however be relatively resilient against database connection issues and will try to reconnect if it fails to connect on startup.