tracing: Disable colored log output on Heroku (#7505)
Having ANSI escape sequences sprinkled across the log lines helps the readability on Papertrail, but Datadog does not support them and displays them as their regular characters which makes the logs hard to read and parse. This commit disables the color support when run on Heroku to address that issue.

This might also cause a small decrease in our log volume due to the escape characters no longer counting against it.