tests/util/test_app: Increase default database connection pool size
Before #7244 we were limited to only having a single database connection in most of our test suite since we were working within a running transaction. This is no longer the case and e.g. our background jobs might use another connection from the pool, while the runner itself already holds a connection with a transaction.

This commit changes the default pool size to 5. Since the connections are opened on demand in the background it should not significantly influence the test performance.