middleware: Move compression and timeout layers out of the `ServiceBuilder` (#7665)
Having `RequestBodyTimeoutLayer` (which changes the `Body` type of the `Request`) in the middle of the `ServiceBuilder` has already been hard to deal with in the past, and straight-up doesn't seem to work anymore with axum 0.7. This change moves these three layers out of the `ServiceBuilder` and applies them to the `Router` directly instead.