Fix `User-Agent` header requirement error responses
The middleware that is responsible for rejecting requests without a `User-Agent` header runs after the middleware that logs requests. The request logging middleware however was expecting a `TypedHeader<UserAgent>`, which meant that requests without such a header were unconditionally rejected, even for download requests.

Our test suite for this functionality was setting an empty header, which unfortunately behaved differently than not having such a header at all.

Related:

- 44aeb47b56037890f49c05464a973b8447f7c0ff implemented the corresponding tests
- 88bd12977523bf68f62d622447c2b9cbbc2f8987 changed the logging middleware to rely on `TypedHeader<UserAgent>`