typosquat/test_util: Remove `conn` field from `Faker` struct
Using `PgConnection` prevents us from using a pooled connection. An alternative would be to store the mutable reference to the connection in the `Faker` struct, but then we have to deal with explicit lifetimes, and the `borrow_conn()` and `into_conn()` fns are still causing a bit of weirdness. This change adjusts the `Faker` API to roughly match the rest of the codebase by accepting an explicit `&mut PgConnection` argument everywhere.