Replace "invalid semver" errors with `version_not_found()`
The `semver::Version::parse()` call is only used to avoid getting a database connection from the pool when we already know that such a version can't exist in the database. In that case we don't need to reply with a special error message though, and can instead reply with the same error that we send when a version isn't found in the database.