controllers/crates: default to sort by recent downloads
"Browse All Crates" is the most prominent link in the menu bar, displayed on all pages. Before this change, this link would present the user with (mostly) irrelevant crates, as it was sorted alphabetically: crates like `a`, `a-`, `a0`, etc.

With this change, the list is now sorted to present the most downloaded crates first.

`/keywords/:term` already defaults to that same order.