serializers/crate-owner-invite: Fix `users` side-loading
Previously, the `normalizeResponse()` implementation was throwing away the `users` array in the payload, because of a "missing id" issue (see https://github.com/rust-lang/crates.io/pull/3490). The root cause for that however was not that the `id` field in the `user` objects was missing (it actually exists), but that the `modelNameFromPayloadKey()` implementation was unconditionally returning `'crate-owner-invite'` even for `user` payloads. This change fixes the `modelNameFromPayloadKey()` implementation to be compatible with `user` payloads and removes the now obsolete `normalizeResponse()` implementation.