worker/jobs/downloads/queue: Delete messages *after* processing them
Previously, if enqueuing the job failed, the message would have been deleted already and the log file would never be processed. With this new implementation if enqueuing the job fails, the message will not get deleted from the queue. The downside of this is that if deletion fails for some reason, we might enqueue the job multiple times. This could already happen though if the code is running concurrently and we will need to introduce a deduplication mechanism anyway.