Move `process_cdn_log` background job to a dedicated worker queue
Running multiple instances of `save_to_version_downloads()` in parallel appears to be causing issues and job retries. This change moves the corresponding background job to a dedicated worker queue with only a single worker. This should prevent these jobs from running in parallel, similar to how we don't run any jobs in parallel that touch the git index repository.