worker/jobs/downloads/update_metadata: Implement optional batch processing
The previous implementation to update the `versions` and `crates` table was quite slow and creating a lot of bloat in the database, since it updated the `crates` and `metadata` tables for every single version that was downloaded.

This commit adds a second implementation, guarded by the `BATCH_UPDATE_DOWNLOADS` feature flag, which updates the tables in batches with a static query that does not need to transfer data back and forth between the background worker and the database.