downloads: Avoid `spawn_blocking()` and database query if `cdn_log_counting_enabled` is set (#8127)
Previously, we were just not calling `downloads_counter.increment()` if the feature flag is enabled. This commit changes the code to not perform the `version_id` database query, not get a database connection from the pool, and not get a blocking thread from tokio via `spawn_blocking()`. This should speed up the download endpoint, and make it resilient to database issues.