database: Use `CONCURRENTLY IF EXISTS` to drop the `index_follows_user_id` index
The `CONCURRENTLY` allows us to still use the associated table while the index is being dropped, and the `IF EXISTS` ensures that the migration succeeds even if the index does not exist. This allows us to manually drop the index before the migrations run, instead of dropping it as part of a migration which would block the server startup.