worker/jobs/downloads/update_metadata: Use `SELECT ... FOR UPDATE` in `batch_update()`
We've seen a couple of deadlock errors in the logs originating from the query in the `batch_update()` function. It looks like in some cases when a `process_cdn_log` background job is running its `UPDATE` query leads to the deadlock observed in the logs. My assumption is that this happens between the initial `SELECT` in the `batch_update()` query, and the `UPDATE` at the end which sets the `counted` column to its new value. By using `FOR UPDATE` we can hopefully prevent the `process_cdn_log` from modifying the rows while the `batch_update()` function is working, which should prevent the deadlock.