components: add a `PrivilegedAction` component
This component wraps controls that can be used by either an owning user
or by an admin. The `@userAuthorised` argument is used to indicate that
the current user is allowed to perform the action wrapped by the
component, while the admin status is extracted from the current session.

If an admin is logged in, but sudo mode is disabled, then a disabled
version of the control(s) will be shown by default. This can be
customised by specifying the `:placeholder` block.