Add `clean_processed_log_files` background job (#8278)
Our `processed_log_files` is already at 35k+ rows and growing steadily. We don't want it to grow infinitely though, so this commit implements a new background job that can be run regularly to remove obsolete rows from the table. The cut off date is currently hardcoded to one week, which should be sufficient to prevent duplicate SQS events and keep the table at a reasonable size.