Implement rate limiting for e-mail verifications.
This applies a burst of 3 and a refill time of 30 minutes by default
per user. (Not that I can imagine a scenario where we'd ever override
this for a user, but using the same machinery as other user actions is
obviously much simpler.)

I don't love that this ends up essentially prop-drilling the rate
limiter into a bunch of new places, but I don't see an alternative other
than prop-drilling the whole app struct through, which would be worse.

This doesn't address (sorry) per-address rate limiting, but is at least
a reasonable starting point.