storage: Use old `http` crate for HTTP header imports
The `object_store` crate unfortunately does not re-export the `http` types that we need for the `with_default_headers()` calls and it is (via `reqwest`) still relying on the v0.2 version of the `http` crate. Since this blocks us from upgrading the `reqwest` dependency ourselves we'll explicitly add the dependency for now until `object_store` is updated to the most recent version of `reqwest` and `http`.