worker/jobs/expiry_notification: Reduce log level for missing email addresses
Ideally, we wouldn't have users without associated email addresses, but for some reason we currently appear to have such cases. Instead of treating this as a hard error for the token expiry notification we can log it with `INFO` level and mark the token as processed even if no notification has been sent out.