Move `MockSentryService` class into the `app` folder (#8704)
The production build is currently broken because the `app-hook` initializer is unable to import the `MockSentryService` class from `crates-io/tests` because those test-only files are not included in the production build. Ideally we'd exclude this initializer from the production build, but I haven't found a straight-forward way for doing that yet. The current solution is not ideal because we are shipping these unnecessary bytes to our users, but until we find a better way this at least unblocks our `main` branch again.