worker/jobs/dump_db: Only export the `public` schema via `pg_dump`
Postgres databases sometimes have additional "schemas" in them, occasionally with different access permissions. One example of this is the `cron` schema that the Scheduler feature on CrunchyBridge creates. This schema is not accessible by the `application`, which causes `pg_dump` to break. By restricting `pg_dump` to the `public` schema we can fix this issue.