worker/jobs/dump_db: Remove `database_url` field (#8836)
On our staging environment we usually don't use a read-replica database, so `READ_ONLY_REPLICA_URL` might not always be set. Instead of saving the potentially secret database URL in the background job we can read it from the background worker config instead, and fall back to the primary database if a read-replica is not configured.