tests: Increase database pool size (#8994)
This was decreased in 933576ae8cf92a1afdcbdafa209eeb73abbe00b2 when we were running the sync and async pools in parallel, but at this point we only use the async pool and can increase the pool size again. This should hopefully reduce the current flakyness of our test suite a bit.