ci: build backend tests in /mnt
GitHub's ubuntu-22.04 runner only guarantees 14 GB of free space on /,
which turns out to not quite be enough any more for our backend tests to
be built and run. GitHub runners also provide a larger temporary
partition on /mnt, so let's use that for our target directory.

This appears to be very slightly slower (~10% slowdown) than building
and running on /, but since it works and / doesn't, here we are.

Note that the presence of /mnt isn't really guaranteed by the GitHub
documentation. It was apparently previously suggested by the Azure
docs[^azure], but the current version doesn't include any reference to
/mnt that I can see.

I don't think there's a lot of downside in us making this change right
now. It feels fragile, but our other option would pretty much be to use
something like the `maximize-build-space` action[^action] to take out
the parts of the ubuntu-22.04 image that we don't need, which is just as
implementation specific. Or we could run tests for each workspace member
in turn, but then we'd have to clean up and would probably lose all the
benefits of the caching we have in CI right now.

Fixes #9050.

[^action]: https://github.com/easimon/maximize-build-space
[^azure]: https://learn.microsoft.com/en-us/previous-versions/azure/jj672979(v=azure.100)?redirectedfrom=MSDN