Make `<source>` elements in `<picture>` adapt to the user's color scheme preference
The `media` attribute of these elements only works with the device preference, but not if the user manually chooses a different color scheme on the website. This Ember.js element modifier moves the original `media` attributes to `data-media` and then overrides `media` with `all` or `none` depending on the passed in color preference. The color preference is taken from the `color-scheme` service, which automatically reacts to the user choices.