Change version downloads archive location to static S3 bucket (#9180)
Previously we were uploading the version downloads exports to a dedicated bucket, but the infra team decided that it would be best to use the `static` bucket instead. This changes the upload location by switching from a dedicated `ObjectStore` instance to a `PrefixStore` based on our `Storage` struct and the `archive/version-downloads` prefix.