jobs/downloads/process_log: Decrease batch size (#9195)
On our production servers we've been seeing warnings like:

> Failed to fill temp_downloads table: error encoding message to server: value too large to transmit

It looks like this error is returned by `tokio-postgres`, which can't handle the amount of data we throw at it when using a batch size of 10k rows. This commit decreases the batch size in the hope that this makes the background job succeed. If not, we might have to further decrease it.

A potentially better solution for this would be to use a `COPY` query, but that wasn't available yet in `diesel` when the code was originally written.