Generate a static index.html for the version download archives
[As suggested by @Turbo87][zulip], we should really have an index here
so people can easily enumerate the version download archives for
whatever nefarious purposes they have.

This will probably need some S3 side configuration as well to serve up
the index.html on demand.

I've set this up as a separate job from `ArchiveVersionDownloads` just
in case we ever need to run it separately, but I don't feel very
strongly about that. It could just as easily be a few new functions in
that module and run as part of that job, instead of being triggered by
that job.

This requires #9206.

[zulip]: https://rust-lang.zulipchat.com/#narrow/stream/318791-t-crates-io/topic/archiving.20.60version_downloads.60.20to.20S3/near/455657391