fix: don't read .psqlrc for automated DB export
The dump_db::run_psql() function spawns a psql child process to run a an
export script, but psql will automatically load and execute the content
of the user's ~/.psqlrc file if it exists and execute it before the
export script.

The statements executed from the psqlrc file might change connection
parameters (for example, the search path) in a way that causes the
script to fail.

This commit stops psql sourcing the psqlrc file at startup, meaning the
connection should always be in a fresh / clean state.