fix: bound co-owner invites per request
Bound the number of co-owner invites that can be sent in a single API
request, effectively bounding the work done per request.

This limits the number of iterations the N^2 membership test loop has to
perform, limiting the amount of CPU time and memory needed for both it
and the subsequent database queries per invitee.