refactor: separate Crate::owner_add model concerns
This separates out the concerns of interacting with the database via the
Crate model, moving management of invites (emailing, response
generation, what is considered success) to the controller.

This will simplify future work on rust-lang/crates.io#2639 by allowing
the controller to inspect the result of the owner_add() call to see if
an invite was generated, or already existed without having to peek at
strings.