CrateSidebar: Adjust "Report Crate" email link
In the current email we only include the name of the crate, but it would be a lot more ergonomic for the responders to be able to directly click a link to the crate. This commit adjusts the email template to use an HTTP link to the crate instead of just the raw crate name.