Use inline snapshots for error responses
The downside of this is that the code needs to be recompiled when the expected error messages change, but it does bring the error messages closer to the test code, which makes the tests a little bit easier to digest. `cargo insta test` makes it fairly easy to update these expectations automatically, even for inline snapshots.