Use more snapshots for error responses
We can hardcode the error messages in our tests but that makes it quite hard to update the error messages, since they have to be updated in multiple places. With insta snapshots they will still have to be updated everywhere but `cargo insta test` can automate this process, which makes it a lot more straight-forward.