Remove obsolete `server_binary` tests
The main purpose of these tests was to check that download requests are switching over to "unconditional redirects", but that mode has been the default for almost a year now and the API endpoint is no longer accessing the database. The secondary purpose was to check how the server behaves in unhealthy database conditions in general, but we already have the `unhealthy_database` tests module for that.

These tests also rely on `env!("CARGO_BIN_EXE_server")`, which means that they have to be integration tests and can not be moved into the `crates_io` library itself. Since these tests rely on a few test utilities this means that all other test relying on these utilities also have to be integration tests to be able to share the code.

tl;dr the tests are no longer particularly useful and enforce too many constraints on their surroundings