Implement `ReleaseTracks`
This will add `release_tracks` information to the meta field for
paginated versions. This is necessary for pagination because it only
loads a portion of versions per page, while `release_tracks` may require
all sorted versions to determine their values. Therefore, we also need
to calculate these on the server side.