Return downloads for the top 5 most downloaded versions
This changes the downloads endpoint to return the five most downloaded
versions of a crate, instead of the five most recent versions. This
should make the endpoint more useful for users who are interested in
seeing the most popular versions of a crate.

In particular it helps with the case where a crate has many pre-release
versions, which are not as interesting to users as the stable versions.