Inline `NewCategory::create_or_update()` (#9715)
The function is only used in our test suite, the update behavior isn't actually needed there and the return value is also not used anywhere. This makes it possible to significantly simplify the function to the point where there is no point anymore to keep the function. Inlining it then also allows us to merge a couple of calls into single queries that insert multiple rows at once.