worker/tests: Fix flaky deduplication test
The second `assert_none!()` in this test was occasionally failing due to a race condition. The `TestJob` that was enqueued right before this one would sometimes already be processed before we enqueue this one, leading to this one being actually enqueued instead of being deduplicated. The reason for this was that the `runner()` fn builds a `Runner` with two workers for the default queue. This commit changes the runner to only use a single worker, which avoids the race condition.