controllers/krate/publish: Reduce `spawn_blocking()` scope (#10003)
Quite a bit of code that is currently within the `spawn_blocking()` scope isn't using the sync database connection, so we might as well move it outside. One notable exception is the `process_tarball()` call, which stays within a dedicated `spawn_blocking()` call for now. While `process_tarball()` is operating on in-memory bytes and should be fairly fast, we sometimes process multiple megabytes of data here, so we can't rely on this being fast enough for an async thread.