database_dump: Fix `num_no_build` import issue
The `num_no_build` column in the database is `NOT NULL`, but since we don't export it in the database dump it will show up as an empty value and thus fail to get imported correctly. This commit temporarily drops the `NOT NULL` constraint during the data import, backfills the data from the `num` column, and then adds the constraint again.