components: don't infer MSRV based on edition (#10108)
In #9996, we started showing the MSRV based on the edition for crate
versions that declare an edition but not a `rust-version` in their
manifest. This isn't strictly correct: a crate may require features
added later in an edition than in the first Rust version that defined
that edition, and showing this as an MSRV implies a level of support
that the crate author may not have intended.

So, we shouldn't show an MSRV if one isn't declared, but we _can_ show
the edition if one of those is defined. Let's do that.

Fixes #10103.