crate: `findRecord` requests for `crate` without including `versions` by default
This will ensure that the `versions` are fetched via a separate
requests, rather than being included in the crate's response. This not
only enables us to render the page layout first for a faster first paint
but also allows us to migrate to a paginated version list in the future.