controllers/krate/metadata: Add support for `default_version` include mode
This allows us to respond with a version data of `default_version`
included, which potentially benefits apps by eliminating the need to
wait for the crate response to obtain the default version and then make
a subsequent request to get the actual version data. This is
particularly useful when we move towards not requesting with `versions`
included.